Updated Files module tests

+++ test/UnitTests/Files/CustomCADs.UnitTests.Files.Application/Cads/SharedCommands/SetCoords/SetCadCoordsData.cs
@@ -1,3 +1,3 @@
 ﻿namespace CustomCADs.UnitTests.Files.Application.Cads.SharedCommands.SetCoords;
 
-public class SetCadCoordsData : TheoryData<int, int, int, int, int, int>;
+public class SetCadCoordsData : TheoryData<decimal, decimal, decimal, decimal, decimal, decimal>;
