+++ test/UnitTests/Files/CustomCADs.UnitTests.Files.Domain/Cads/Create/WithId/CadCreateWithIdData.cs
@@ -1,3 +1,3 @@
 ﻿namespace CustomCADs.UnitTests.Files.Domain.Cads.Create.WithId;
 
-public class CadCreateWithIdData : TheoryData<string, string, int, int, int>;
+public class CadCreateWithIdData : TheoryData<string, string, decimal, decimal, decimal>;
